Теперь рендеринг страницы осуществляется с помощью контроллера page-controller.js
